Consolidate CSS selector logic to one shared file
