+++ pyspelling/util/__init__.py
